Fixes 3075: Stop using reflection for ITemplateEngineHost.BuiltInComponents
# Problem
Currently its IReadOnlyList<KeyValuePair<Guid, Func<Type>>> which in practice means that AssemblyComponentCatalog needs to do assembly.GetTypes() and then do bunch of reflection to figure out which interfaces are implemented, and on top of all that we also call Activator.CreateInstance(type).

What we want to do is change that property type to IReadOnlyList<(Type, IIdentifiedComponent)>, which should save ~100ms on every run of dotnet new

# Solution
Each assembly that is part of our repo and has components now has in root namespace "Components" class that returns all components and unit tests that validates that is true.
Now host can filter out some component and whole thing should be faster and give us more flexibility when writing unit tests since we control better instances and their values in tests vs. passing in type where we had to set some static values which is not best when unit tests execute in parallel.
Also moved SettingsStore inside ComponentManager since it was only user of it...